preparing for content additions
