+++ src/components/layout/DevHomework.js
@@ -4,7 +4,7 @@ import logo from '../img/ff-logo.png';
 
 export default ({ children }) => (
   <div className="homework">
-    <div className="contain">
+    <div className="container">
       <img className="homework__logo" alt="Firefox Logo" src={logo} />
       <h2 className="homework__title">Firefox DevTools + CSS Grid Layout</h2>
       <div className="row">
