+++ src/data/css-grid.js
@@ -2,52 +2,52 @@ const pageData = [
   {
     order: 1,
     title: 'Terminology',
-    link: '/',
+    link: '/css-grid',
   },
   {
     order: 2,
     title: 'Your First Grid',
-    link: '/02-first-grid/',
+    link: '/css-grid/02-first-grid/',
   },
   {
     order: 3,
     title: 'Firefox DevTools',
-    link: '/03-firefox-devtools',
+    link: '/css-grid/03-firefox-devtools',
   },
   {
     order: 4,
     title: 'The fr Unit',
-    link: '/04-fr-unit',
+    link: '/css-grid/04-fr-unit',
   },
   {
     order: 5,
     title: 'Mixing Units',
-    link: '/05-mixing-units',
+    link: '/css-grid/05-mixing-units',
   },
   {
     order: 6,
     title: 'Position Items',
-    link: '/06-position-items',
+    link: '/css-grid/06-position-items',
   },
   {
     order: 7,
     title: 'Basic Layout',
-    link: '/07-basic-layout',
+    link: '/css-grid/07-basic-layout',
   },
   {
     order: 8,
     title: 'Template Areas',
-    link: '/08-template-areas',
+    link: '/css-grid/08-template-areas',
   },
   {
     order: 9,
     title: 'Named Lines',
-    link: '/09-named-lines',
+    link: '/css-grid/09-named-lines',
   },
   {
     order: 10,
     title: 'Learn More',
-    link: '/10-learn-more',
+    link: '/css-grid/10-learn-more',
   },
 ];
