+++ src/pages/css-grid/02-first-grid.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Main from '../components/layout/Main';
-import CodeBlock from '../components/CodeBlock';
-import DevHomework from '../components/layout/DevHomework';
-import CodepenLink from '../components/CodepenLink';
+import Main from './components/_Main';
+import CodeBlock from '../../components/CodeBlock';
+import DevHomework from '../../components/layout/DevHomework';
+import CodepenLink from '../../components/CodepenLink';
 
 const Tutorial = () => (
   <div>
